Move fen and most of hash into main Board impl, change some hash details
With the BoardBuilder handling creating boards with fens, the fen module
didn't really need to exist anymore.
Also moved the Board impl in hash to the main impl in Board so most
Board methods are now in the same place. Ideally I want all methods to
be in the same impl.

Also changed how the hash gets updated when moves are made. I added
methods for all updates that are tracked by hash to the Updater struct
so that all hash updates are now done through that class. To check that
everything is still correct with hashing, I added the
collision_detection test.