Add move cache to board and update on every piece change
The move cache will speed up the move cycle (move generation -> make
move) by combining the work done to check for move legality (keeping a
cache of possible attacks by the opposing team) and move generation into
a single, incremental process. It should provide a speed boost to both.

This commit only adds the incremental process of generating the move
cache during each turn. It does not use the cache to do anything like
check for the legality of moves or generate possible moves, so it is a
performance regression.

I also haven't tested that the generated cache is correct in any way
yet.